Rewrote some code in the controller and added mixins for flexbox
The 2.1.x versions will be all about rewrites. I still need to separate some code and perhaps use a factory for the player instead of the controller. Next version will also implement the ui-router for separating the homepage.

The new flexbox mixins will keep the scss a bit cleaner and structured.
The new rewrite of the homecontroller allows for a more cautious approach.