Works again, but still needs an overhaul. This code makes me itch again. Mobile also plays now, but won't with external push, but we'll get to that later. At least it runs locally on the device.
Also i'm trying to find out what i want to do with this project. Seems like a nice easy and simple app but not sure what road to take. Just mobile, add LAN support (or whatever), add custom sounds, or something like that.

In the overhaul this will be dealt with, same with cleaner code (and less dependencies), better design (especially on mobile) and beter local-hosting.